vim: Drop Neocomplete
It's cool and works fine (expect on Windows network drives where it
slows down Vim as hell), but I don't use the capabilities of it.

The omnicomplete is better for Ruby and Jedi is far enough for Python,
Necocalc is unnecessary since I using Scriptease's `g!`.

Where Neocomplete would be better than the default omnicomplete is the
Vim filetype, but it makes no sense since I called the completion
manually.