git: Stash keeps staged files
While stashing, all of the files will dissapear from the working tree,
which makes hard to do tests before commiting. For example if A, B and C
files are modified and I'm not sure that A works fine.

With the `--keep-index` flag it's easy to test: stage A,
`stash --keep-index` (which will stash the unstaged files), test A,
commit the changes of A, `stash pop` and start over with B.