ahk: WindowPadX is sometimes buggy
For example if the "Save as" is opened and I start the typing of the
filenam which begans with numbers, then it resizes the window instead of
typing.

Besides this I realized that I don't need but the left/right/maximize
modifications (or real i3 on Windows) so added hotkeys to those.