git: Add aliases
The `git sh` is useful on Windows to open Git Bash in the current
directory. The test on `cd` is neede, because if the command is used
inside a repository, then the shell is opened in the root of it and the
`pwd` does not works for some reason.

Use `git las` to overview the topology of branches.