thunderbird: Search by `Message-ID`
The `Message-ID` is shown in my commits (if necessary) and I need to
search for it in Thunderbird.

To get an e-mail's message-ID, go to `Tools` -> `Options`, select the
`Advanced` panel, select the `General` tab, and click `Config Editor`
and set `mailnews.headers.showMessageId` to `true`.

To search for it, go to `Edit` -> `Find` -> `Search Messages` (or press
`Ctr-Shift-F`). From the search field select the last option `Customize
...`, in the text box of the new apperead window write `Message-ID` and
click `Add` and then `OK`. Now you have a new field to search by, select
it write your ID in the search field.