!vim!neobundle: Hardcoded conditions for plugins
Until now I used the `disabled` setting of NeoBundle to disable plugins
conditionally, but some plugins (like UltiSnips) adds autocommands which
cannot be executed if the plugin is not loaded.

For example UltiSnips adds a `FileType` autocommand, because NeoBundle
evaulates the 'ftdetect' directory, but the plugin itself is not
enabled, so the autocommand (which calls an UltiSnips autoload commands)
cannot be executed:

  Error detected while processing FileType Auto commands for "*":
  E117: Unknown function: UltiSnips#FileTypeChanged

EDIT:
Just found out that deleting ~/.vim/bundle/.neobundle/ftedect/ftdetect
solves the problem.