!vim!dein: Move unnecessary commands out from Dein_...
The plugin initializer functions contained unnecessary commands which
slowed down the startup. Moving those out of the functions.

                              | old     | new
Without file                  | ~590ms  | ~720ms
With file                     | ~950ms  | ~950ms
With file and ~/.vimrc_viewer | ~870ms  | ~950ms

The results are bad, but it seems to be faster (does not waiting after
Vim is opened).

____________________________________
Keywords: !startuptime