!git: Use `--force-with-lease` for push
This option allows you to say that you expect the history you are
updating is what you rebased and want to replace. If the remote ref
still points at the commit you specified, you can be sure that no other
people did anything to the ref. It is like taking a "lease" on the ref
without explicitly locking it, and the remote ref is updated only if the
"lease" is still valid.