!vim_neocomplete!vim_mucomplete: Switch plugin
Mucomplete is async and uses native completions (omnicomplete, keyword
complete, etc.).

Neocomplete on the other hand is a bit clever in certain cases (for
example on Vim scripts) but it's slower and does not opening preview
window when completing.

When Neovim becomes stable, I will use Deoplete - until that I using
Mucomplete.