Revert "!vim: Extract statusline to a function"
This reverts commit bd51055e9292616756ac81c2fbae1402bfdea94c.

Because of `%!` the function is evaulated in the context of the current
window instead of the statusline's (even inactive) window, thus the
filetype for example are the same for each window.

Using `%{StatusLine}` does not parsing hlgroups so it would need to
split the function in to several parts and add the coloring in the `set
statusline`.