!linux!kde: Move config files to .local
It seems that a lot of rc and config files moved to ~/.local. \o/

I did minimal fixes, tweaks besides moving the files.

It seems that there was some configs which is too heavy to check
manually, I will set up the things again and add only relevant
lines to Git.