!git: Drop alias of amend commit
Do not amend commits: if you working on a branch and you are sure that
there will be only one commit, but you doing this by amending, stop it
immediately! If you make a mistake in the the code but you don't
realizing it, it will smash up the only one commit. Always add new
commits, in the end you can squash those in to one without losing
anything. Besides this you can compare the changes (in speed, behaviour,
whatever) between commits.

Besides this I droped the alias of `git commit -m` because it does not
fires up prepare-commit-message hook (add issue tracker number by branch name
for example).