!git!meld: Use Meld as difftool
Vim is very good in editing files, but the diff is not as usefull as
Meld: the latter has syntax highlighting and able to diff dirs.
Unfortunatelly it cannot compare images.

I tried out P4Merge too, it can diff images, but it doesn't have syntax
highlighting.

Kitty' (the terminal emulator) builtin diff is really rocks, maybe I
should rewrite my config to use it on Linux, and Meld on Windows (by
storing the difftool program's name in a variable, `GIT_DIFFTOOL` for
example).