!git: Refine commit template
I just realized that I resting my eyes on the comments when I stop
writing, thus I want to see all of the comments. I'm using `cc` in Vim
to delete the first line and enter in insert mode - it's the "WHY?"
line, thus I cannot read it after deleting.