!bash!git: More verbose Git prompt
For exampe:

  master *$<> origin

*      dirty
< > =  before / after upstream
$      stash

I had to play with sed, because there is a space before the stash ($),
but there is no space before the other characters and I wanted one space
on each side.

I cut the branch name from upstream (origin/master -> origin) because it
takes up a lot of space when I use long branch names.