!git: Interpret the issue number in branch name even without a brief summary
So far, it has only found the issue number "123" in the
"GITHUB-123-fix-typo" branch if the issue number was followed by the
brief summary.

Since we may not want to specify this (because we only named the branch
specifically to have the issue number in the commit message), the brief
summary has to be optional.